requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.0
ddgs==9.13.0
playwright>=1.44.0
